+++ regional_tracking/__init__.py
@@ -1,9 +1,9 @@
-from regional_tracking.raw_object import *
-from regional_tracking.tracking_object import *
-from regional_tracking.line_interset_util import *
 from regional_tracking.area_tracking import *
+from regional_tracking.feature_vector import *
+from regional_tracking.gfx_utils import *
 from regional_tracking.line_boundary_check import *
+from regional_tracking.line_interset_util import *
+from regional_tracking.raw_object import *
 from regional_tracking.regional_detect_tracker import *
 from regional_tracking.regional_detect_tracker_2 import *
-from regional_tracking.gfx_utils import *
-from regional_tracking.feature_vector import *
+from regional_tracking.tracking_object import *
