+++ regional_tracking/area_tracking.py
@@ -1,20 +1,3 @@
-from audioop import cross
-import cv2
-import numpy as np
-
-from regional_tracking import Point, line_intersect
-from regional_tracking.line_boundary_check import point_polygon_test, calc_vector_angle
-
-
-# ------------------------------------
-# Area intrusion detection
-
-class Area:
-    def __init__(self, contour):
-        self.contour = np.array(contour, dtype=np.int32)
-        self.count = 0
-
-
 # Area intrusion check
 def checkAreaIntrusion(areas, objects):
     # global audio_enable_flag
@@ -26,51 +9,24 @@ def checkAreaIntrusion(areas, objects):
                 area.count += 1
 
 
-# Draw areas (polygons)
-def drawAreas(img, areas):
-    for area in areas:
-        color = (0, 0, 255) if area.count > 0 else (255, 0, 0)
-        cv2.polylines(img, [area.contour], True, color, 4)
-        cv2.putText(img, str(area.count), (area.contour[0][0], area.contour[0][1]), cv2.FONT_HERSHEY_PLAIN, 4, color, 2)
-
-
-# in: boundary_line = boundaryLine class object
-#     trajectory   = (x1, y1, x2, y2)
-def checkLineCross(boundary_line, trajectory) -> bool:
-    traj_p0 = trajectory[0]  # Trajectory of an object
-    traj_p1 = trajectory[1]
-    bLine_p0 = boundary_line.p0 # Point(boundary_line.p0[0], boundary_line.p0[1])  # Boundary line
-    bLine_p1 = boundary_line.p1 # Point(boundary_line.p1[0], boundary_line.p1[1])
-    intersect = line_intersect(traj_p0, traj_p1, bLine_p0, bLine_p1)  # Check if intersect or not
-
-    if intersect is True:
-        boundary_line.setIntersect(flag=intersect)
-        angle = calc_vector_angle((traj_p0.x, traj_p0.y),
-                                  (traj_p1.x, traj_p1.y),
-                                  (bLine_p0.x, bLine_p0.y),
-                                  (bLine_p1.x, bLine_p1.y))  # Calculate angle between trajectory and boundary line
-        if angle < 180:
-            boundary_line.count1 += 1
-        else:
-            boundary_line.count2 += 1
-        # cx, cy = calcIntersectPoint(traj_p0, traj_p1, bLine_p0, bLine_p1) # Calculate the intersect coordination
-
-    return intersect
-
-# Multiple lines cross check
-def checkLineCrosses(boundaryLines, objects):
-    objs_with_trajs = [obj for obj in objects if len(obj.trajectory) > 1]
-    for obj in objs_with_trajs:
-        lines_not_crossed = [bline for bline in boundaryLines if bline.uuid not in obj.crossed_lines]
-        for line in lines_not_crossed:
-            for idx in range(len(obj.trajectory) - 1):
-                p0 = Point.point(obj.trajectory[idx])
-                p1 = Point.point(obj.trajectory[idx + 1])
-                if checkLineCross(line, [p0, p1]):
-                    obj.crossed_lines.append(line.uuid)
-                    break
-
-
-def resetLineCrosses(boundaryLines):
-    for line in boundaryLines:
-        line.resetIntersect()
+# Test whether the test_point is in the polygon or not - 指定の点がポリゴン内に含まれるかどうかを判定
+# test_point = (x,y)
+# polygon = collection of points  [ (x0,y0), (x1,y1), (x2,y2) ... ]
+def point_polygon_test(polygon, test_point) -> bool:
+    if len(polygon) < 3:
+        return False
+    prev_point = polygon[-1]  # Use the last point as the starting point to close the polygon
+    line_count = 0
+    for point in polygon:
+        # Check if Y coordinate of the test point is in range
+        if min(prev_point[1], point[1]) <= test_point[1] <= max(prev_point[1], point[1]):
+            # delta_x / delta_y
+            gradient = (point[0] - prev_point[0]) / (point[1] - prev_point[1])
+            # Calculate X coordinate of a line
+            line_x = prev_point[0] + (test_point[1] - prev_point[1]) * gradient
+            if line_x < test_point[0]:
+                line_count += 1
+        prev_point = point
+    # Check how many lines exist on the left to the test_point
+    included = True if line_count % 2 == 1 else False
+    return included
