+++ regional_tracking/line_boundary_check.py
@@ -1,41 +1,16 @@
-import uuid
-
 import numpy as np
 from numpy import linalg as LA
 
-from regional_tracking.line_interset_util import Point
-
-
-class BoundaryLine:
-    def __init__(self, line=(0, 0, 0, 0)):
-        self.uuid = uuid.uuid4()
-        self.p0 = Point(line[0], line[1])
-        self.p1 = Point(line[2], line[3])
-        self.color = (0, 255, 255)
-        self.font_size = 4
-        self.line_thickness = 2
-        self.text_thickness = 2
-        self.text_color = (0, 255, 255)
-        self.count1 = 0
-        self.count2 = 0
-
-    def setIntersect(self, flag: bool):
-        self.color = (0, 0, 255) if flag else (0, 255, 255)
-
-    def resetIntersect(self):
-        self.setIntersect(flag=False)
+from regional_tracking import line_intersect, Point
 
 
 # ---------------------------------------------
 # Checking boundary line crossing detection
-
 # def line(p1, p2):
 #     A = (p1[1] - p2[1])
 #     B = (p2[0] - p1[0])
 #     C = (p1[0] * p2[1] - p2[0] * p1[1])
 #     return A, B, -C
-
-
 # Calcuate the coordination of intersect point of line segments - 線分同士が交差する座標を計算
 # def calcIntersectPoint(line1p1, line1p2, line2p1, line2p2):
 #     L1 = line(line1p1, line1p2)
@@ -46,8 +21,6 @@ class BoundaryLine:
 #     x = Dx / D
 #     y = Dy / D
 #     return x, y
-
-
 # Check if line segments intersect - 判断线段是否相交，这个算法是错误的。当一个点在另外一个条线段上时，该算法有问题
 # def checkIntersect(p1, p2, p3, p4) -> bool:
 #     tc1 = (p1[0] - p2[0]) * (p3[1] - p1[1]) + (p1[1] - p2[1]) * (p1[0] - p3[0])
@@ -75,30 +48,49 @@ def calc_vector_angle(point1, point2, point3, point4):
     n = LA.norm(u) * LA.norm(v)
     c = i / n
     a = np.rad2deg(np.arccos(np.clip(c, -1.0, 1.0)))
-    if u[0] * v[1] - u[1] * v[0] < 0:
-        return a
-    else:
-        return 360 - a
+    return a \
+        if u[0] * v[1] - u[1] * v[0] < 0 \
+        else 360 - a
 
 
-# Test whether the test_point is in the polygon or not - 指定の点がポリゴン内に含まれるかどうかを判定
-# test_point = (x,y)
-# polygon = collection of points  [ (x0,y0), (x1,y1), (x2,y2) ... ]
-def point_polygon_test(polygon, test_point) -> bool:
-    if len(polygon) < 3:
-        return False
-    prev_point = polygon[-1]  # Use the last point as the starting point to close the polygon
-    line_count = 0
-    for point in polygon:
-        # Check if Y coordinate of the test point is in range
-        if min(prev_point[1], point[1]) <= test_point[1] <= max(prev_point[1], point[1]):
-            # delta_x / delta_y
-            gradient = (point[0] - prev_point[0]) / (point[1] - prev_point[1])
-            # Calculate X coordinate of a line
-            line_x = prev_point[0] + (test_point[1] - prev_point[1]) * gradient
-            if line_x < test_point[0]:
-                line_count += 1
-        prev_point = point
-    # Check how many lines exist on the left to the test_point
-    included = True if line_count % 2 == 1 else False
-    return included
+# in: boundary_line = boundaryLine class object
+#     trajectory   = (x1, y1, x2, y2)
+def checkLineCross(boundary_line, trajectory) -> bool:
+    traj_p0 = trajectory[0]  # Trajectory of an object
+    traj_p1 = trajectory[1]
+    bLine_p0 = boundary_line.p0  # Point(boundary_line.p0[0], boundary_line.p0[1])  # Boundary line
+    bLine_p1 = boundary_line.p1  # Point(boundary_line.p1[0], boundary_line.p1[1])
+    intersect = line_intersect(traj_p0, traj_p1, bLine_p0, bLine_p1)  # Check if intersect or not
+
+    if intersect is True:
+        boundary_line.setIntersect(flag=intersect)
+        angle = calc_vector_angle((traj_p0.x, traj_p0.y),
+                                  (traj_p1.x, traj_p1.y),
+                                  (bLine_p0.x, bLine_p0.y),
+                                  (bLine_p1.x, bLine_p1.y))  # Calculate angle between trajectory and boundary line
+        if angle < 180:
+            boundary_line.count1 += 1
+        else:
+            boundary_line.count2 += 1
+        # cx, cy = calcIntersectPoint(traj_p0, traj_p1, bLine_p0, bLine_p1) # Calculate the intersect coordination
+
+    return intersect
+
+
+# Multiple lines cross check
+def checkLineCrosses(boundaryLines, objects):
+    objs_with_trajs = [obj for obj in objects if len(obj.trajectory) > 1]
+    for obj in objs_with_trajs:
+        lines_not_crossed = [bline for bline in boundaryLines if bline.uuid not in obj.crossed_lines]
+        for line in lines_not_crossed:
+            for idx in range(len(obj.trajectory) - 1):
+                p0 = Point.point(obj.trajectory[idx])
+                p1 = Point.point(obj.trajectory[idx + 1])
+                if checkLineCross(line, [p0, p1]):
+                    obj.crossed_lines.append(line.uuid)
+                    break
+
+
+def resetLineCrosses(boundaryLines):
+    for line in boundaryLines:
+        line.resetIntersect()
