+++ regional_tracking/raw_object.py
@@ -1,6 +1,11 @@
 # ------------------------------------
 # Object tracking
 import time
+import uuid
+
+import numpy as np
+
+from regional_tracking import Point
 
 
 class RawObject:
@@ -41,3 +46,31 @@ class RawObject:
     @property
     def height(self) -> int:
         return self.bottom - self.top
+
+
+class BoundaryLine:
+    def __init__(self, line=(0, 0, 0, 0)):
+        self.uuid = uuid.uuid4()
+        self.p0 = Point(line[0], line[1])
+        self.p1 = Point(line[2], line[3])
+        self.color = (0, 255, 255)
+        self.line_thickness = 2
+        self.textColor = (0, 255, 255)
+        self.textSize = 4
+        self.text_thickness = 2
+        self.count1 = 0
+        self.count2 = 0
+
+    def setIntersect(self, flag: bool):
+        self.color = (0, 0, 255) if flag else (0, 255, 255)
+
+    def resetIntersect(self):
+        self.setIntersect(flag=False)
+
+
+# ------------------------------------
+# Area intrusion detection
+class Area:
+    def __init__(self, contour):
+        self.contour = np.array(contour, dtype=np.int32)
+        self.count = 0
